component does not remount when clicking buttons in the editor toolbar
